+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased] - 2019-01-21
+### Fixed
+ - All arguments are now correctly passed to winetricks.
+
 ## [1.1.1] - 2019-01-20
 ### Added
  - Game-specific Proton installations are now detected.